worker/requirements.txt
aniso8601==9.0.1
click==8.0.3
colorama==0.4.4
coverage==5.5
Faker==14.0.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT==0.2.0
Flask-JWT-Extended==4.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3.post0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
PyJWT==2.1.0
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.15
Werkzeug==2.0.1
gunicorn==20.1.0
jsonschema==4.16.0
celery==5.2.7
redis==3.5.3
python-dotenv==0.21.0
Flask-Mail==0.9.1
psycopg2-binary==2.9.4
google==3.0.0
google-cloud-storage==2.6.0
py-healthcheck==1.10.1
google-cloud-pubsub==2.13.11